git-svn-id: http://svn.linuxonly.nl/public/checkmate/trunk@48 f5a0e3f6-4f0a-0410-a315-d1dc7e33e6ef

+++ mpck/bufio.h
@@ -19,7 +19,7 @@ typedef struct {
 	FILE * fp;
 	char * buffer;
 	char * bufpnt;
-	int buflen;
+	size_t buflen;
 	size_t filesize;
 } CFILE;
